+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ### 5.3.0 (XXXX-XX-XX)
 **SDK**
 - **Deprecated:** `getUniqueId(completionHandler:)` in `MiniAppMessageDelegate` protocol is deprecated. You should use `getMessagingUniqueId(completionHandler:)` instead.
+- **Feature:** Added new `HostAppThemeColors` model class to hold the primary and secondary color strings in Hex format.
+- **Feature:** Added a new optional interface `getHostAppThemeColors(completionHandler:)` in `MiniAppMessageDelegate` protocol, to fetch the host app theme colors.
+
+**Sample App**
+- **Feature:** Implemented method `getHostAppThemeColors(completionHandler:)` in the sample app to fetch the host app theme colors and pass it to MiniApps.
 
 ### 5.2.0 (2023-03-30)
 **SDK**